requirements.txt
Django==4.1.1
whitenoise==6.4.0
djangorestframework~=3.14.0
django-cors-headers=3.14.0
